Initial commit of 6 files, each with empty functions.
Each of the functions that are to be developed as a part of this
project are initialized and split into different groups based on
their functionality.
`mailinglist.py` has two functions, both
of which deal with the email information in the spreadsheet.
`material.py` has two functions that create physical materials
based on the information in the spreadsheet.
`newmember.py` has two functions that deal with adding new
members to the spreadsheet and creating the unique ID.
`querymember.py` includes a function to search the
spreadsheet for member information.
`updatemember.py` includes three functions, updating
member information or moving a member to and from the
inactive member sheet.
`vopmembership.py` is the main function, prompting the user for
what to do and then calling all other functions.