Created two new functions that make two dictionaries with the data from the membership spreadsheet. Added classes.py and mapping.py.
Using `load_workbook` from `openpyxl` and python's built-in
Data Class, the `create_active_dicitonary()` and
`create_inactive_dictionary()` functions make dictionaries of
all the rows in the membership spreadsheet, separating by active
and inactive members.

The dictionary uses member id as the key, and the value is an instance
of the Data class defined in the new file, classes.py. Each of the
columns in the spreadsheet map to an attribute of the data class
instance, and the attribute is assigned the value of the
corresponding cell in the spreadsheet.
 Each attribute is searchable using `dictionary[member_id].attribute`

The file mapping.py maps the column numbers to values, for better
readability of the main functions.