Changed functions so now they output lists instead of dictionaries.
The functions `create_active_dictionary` and `create_inactive_dictionary`
are changed to `create_active_members` and `create_inactive_members`.
The functions now return a list of objects of the Member class,
instead of a dictionary whose values are a list of Member objects.
This is an experiment to see if it simplifies the way one iterates
over the Member objects and calls their attributes. I found that I
wasn't using the key in the dictionary at all, and that making a
dict was unneccessary.