Filled in the `query_member()` function.
`query_member()` searches the list of Member objects for a certain
attribute. It takes two parameters, the filename and the attribute
you wish to search for. It returns a list of all of the attribute
values from the search.

I'm thinking to use this function to help create the member ID,
possibly having it iterate through this list to find all the
ID numbers that already exist and creating a new number that is
in sequence.