Filled in `create_member_id()` function.
`create_member_id()` calls `query_member` functions to create a list
of existing member ids. Iterates over the ids and generates a new
unique id that is sequential. I'll admit that I'm not sure the
loops are well-written, but they do function with the expected
output. I want to come back to the loop in the `create_member_id()`
fucntion and see if I can clean it up and figure out how to get
the results I'm looking for with something that makes more sense.

Also changed the xl file to use member id's that are 1, 2, 3
instead of 001, 002, 003 because python doesn't see those
as int, and it's visually confusing. I also changed the class
attribute to make id a str instead of int.

deleted the test function at the bottom of querymember.py because
it was confusing my output for some reason.