Fixed `create_member_id()` function.
The `create_member_id()` function now searches through a list of all
the current member ids and adds 1 to the largest id to create
a new id. This is much more effecient and generally makes more
sense than how it was before.

Added more comments to classes.py for better readability.