Created the `new_member_email()` function.
new_member_email() takes an existing template, the email of the
recipient, and the spreadsheet from which to read the data as
arguments. It fills out the email template with the information
about the member using the recipient's email to look up the
member's information in the spreadsheet. It uses query_member_attr()
to look up both the new  member's information and the information for
the member's section leader to fill in the template.

It also returns a subject line for the email, containing the
member's first name.

I'm including the new member email template in this commit.

I also updated the README with more sources, and changed something
minor in project.py.