Added 3 more email generation functions.
Added 3 functions that generate emails from a template for updating
member info, moving a member from Inactive to Active, and moving
a member from Active to Inactive. All functions take 3 parameters,
the template from which to generate the email, the email of the
recipient, and the spreadsheet from which to get member information.

Also adding the txt files for the tepmlates.

Updated the documentation and added comments for funcitons in
newmember.py.