Started filling in `main()`.
Created the basic structure for the main() function, including
adding a main menu and creating a loop to prompt the user for
input to do various tasks in the program.

Also added the function add_member() to project.py to better
meet the criteria for the project, which states I must have three
additional functions in my project.py file. I will be adding
all the functions that main() uses to project.py, and leaving
the secondary functions (that my functions use) in files on their
own.

Commented out the moved function in newmember.py because I'm
scared to delete it.

Also added some error handling in add_member().