Modified `generate_email()` function to make it reuasable.
Changed the template files to include the subject line at the
top of the file. Changed `new_member_email()` function to
take the subject line from the file template and put it into
the subject variable, and put the rest of the template into
the body variable. This way I can use the same function
to fill out the four different templates that I have, instead
of using a separate function for each template.

Updated the templates and project.py to reflect the change in
the funciton name.