experimental. Trying to mess with a mailing list function.
Leaving this here for now, because this is posisbly beyond the scope
of the project. Attempting to build a mailing list function
that can send an email to a certain specified group of memebrs
in the active members sheet. could send to all emails, just section
leaders, just people with roles, or just people in a certain section.