Created `inactive_member()` and `active_member()` functions.
These two functions move members between the active and inactive
member sheets on the spreadsheet. They take three parameters -
the file to be modified, the name of the attribute of the member they
want to move, and the value of that attribute.
Looks up that member, then finds them in the spreadsheet and copies
their information from the sheet they are on to the other one.

Also adds the date modified, as well as automatically sends an email
using the email templates when the task completes.

Also created a new query_member function for this. I feel like I have way
too many different query member functions, and I think i could rewrite some
of my other functions to work with a more reusable/universal query function.