Updated project.py with `(in)active_member()` functions.
Added the active_member() and inactive_member() functions
to project.py and commented them out of updatemember.py. Updated
main() with the two new functions, adding to the main() function
the ability to move members back and fourth from the two
sheets of the spreadsheet.

Updated the README with more sources and made minor edits to emails.py.