Created the `search_member()` function.
The search_member() function uses query_member_object() to search
the spreadsheet for members that match the search criteria. Prints
all of the information for the matching members to the screen. If
the search yields multiple members, the user is informed of this
and the function counts each of the results as they are printed
to the screen.

Added this function to main(). In main, choosing the "search member"
main menu item prompts the user to choose from a menu of search
criteria. Based on the criteria given, the main() function uses the
search_member() function to search the spreadsheet using the
attribute given and the attribute value provided.