Created `update_member()` function and added it to `main()`.
The update_member function takes the email of the member you
wish to update, finds that member in the spreadsheet, and prompts
the user to update any of the available fields in the spreadsheet.
Once the user is done updating all the desired fields, it sends
an email to the member whose info was updated summarizing the
updates.

Added update_member to main, filling in another main menu
item. The user is prompted for input of the email, and
then it calls the update_member function to complete
the task.

Also modified the generate_email function to include a field for
"role" because I apparently forgot that field existed. Also had to
update a few templates  to include that field.