Added a new column to the spreadsheet and other minor error handling changes.
Added a "status" column to the spreadsheet that includes information on
whether the member is from the active sheet or inactive sheet. This
also involved adding another attribute to the member class, as well
as additional editing to all the functions that query the member class
for information. This includes adding a "status" line to the search_member
function so that the user can see whether the member they searched for
are active or inactive. Also changed the (in)active_member functions
to include editing the status column to reflect the change in status
when the member is moved from one sheet to another.

Added error handling, particularly at the beginning of main() to check if
the file passed into the function was the right format for all the
following functions, returning an error if it is not the right format.
This will prevent any KeyErrors or file formatting errors from occuring when
the user uses the application.

Added additional error handling in other functions. Fixed an issue where
the new member ID was an int rather than a str (making it unsearchable).