Removed material creation from main() and finalized error handling.
Finally bit the bullet and admitted to myself that creating functions
to make labels and nametags was beyond the time constraints of this
project. Edited the README to reflect this change. Removed the options
from the main menu in main() and deleted the file where those functions
would go.

Also finalized error hadnling, including changing generate_email() to prevent
a badly formatted template from being used and causing an error in creating the
email. Updated the .gitignore to ignore the test template used.