Added `test_project.py` and began unit test building.
Completed unit test building for `add_member()` function in
project.py. The test checks for the input to make sure it appends properly,
checks that the date modified is updated correctly, and checks that the send email
functions generate and send the correct subject, body, and emails.

Also tweaked the aesthetics of the command line interface for the
program, adding emojis and line breaks to make the program more
readable and interesting.